Added a function to ouput tei:date nodes
Input of funcDateFormatTei is the usual string. The output is a tei:date node with @when or @when-custom, @calendar, and @datingMethod attributes depending on the selected calendar. The content of the tei:date node can be toggled between the original input string and a formatted output of "d MNn yyyy".